some updates

+++ src/TiredEagle.Domain/TiredEagle.Domain/BusinessOwner/ValueObjects/Budget.cs
@@ -4,11 +4,11 @@ namespace TiredEagle.Domain.CampignOwner.ValueObjects
 {
     public class Budget : ValueObject
     {
-        public Budget(decimal amount)
+        public Budget(double amount)
         {
             if (amount <= 0) throw new ArgumentException("Budget amount must be positive.");
         }
-        public decimal Amount { get; set; } // Money value object
+        public double Amount { get; set; } // Money value object
         //public string Currency { get; set; }
         protected override IEnumerable<object> GetEqualityComponents()
         {
