+++ src/TiredEagle.Domain/TiredEagle.Domain/ChannelOwner/ValueObjects/ChannelRevenue.cs
@@ -4,7 +4,7 @@ namespace TiredEagle.Domain.ChannelOwner.ValueObjects
 {
     public class ChannelRevenue : ValueObject
     {
-        public decimal Revenue { get; set; }
+        public double Revenue { get; set; }
         public string Currency { get; set; }
         protected override IEnumerable<object> GetEqualityComponents()
         {
